Create script for building gpsim on windows
It's a giant pain but at least with this new python script, many of
the headaches are out of the way.  You still need to install a 32-bit
version of python and 32-bit version of cygwin with specific packages
in order to compile everything.  It works if you do it all right though